remove splash screen
