+++ app/src/main/res/layout/activity_on_boarding.xml
@@ -37,7 +37,7 @@
             android:layout_width="match_parent"
             android:layout_height="50dp"
             android:background="@color/colorPrimary"
-            android:text="Next"
+            android:text="@string/next"
             android:textAllCaps="false"
             android:textColor="@color/md_white_1000"
             android:textSize="18sp"
