--- app/src/main/res/layout/activity_splash.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/colorPrimary"
-    tools:context=".ui.SplashActivity">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:gravity="center"
-        android:orientation="vertical">
-
-        <ImageView
-            android:layout_width="@dimen/user_image_size"
-            android:layout_height="@dimen/user_image_size"
-            android:src="@drawable/ole_logo"
-            app:tint="@color/bg_white" />
-
-        <Button
-            android:id="@+id/get_started"
-            style="@style/BlueButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="@dimen/_10dp"
-            android:background="@drawable/buttonblue"
-            android:padding="@dimen/padding_normal"
-            android:text="@string/get_started"
-            android:textColor="@color/md_white_1000" />
-
-        <TextView
-            android:id="@+id/tv_available_space"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="@dimen/text_margin"
-            android:textColor="@color/bg_white"
-            tools:text="Available Space: 10 GB" />
-    </LinearLayout>
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
